refactor(app): extract CampaignRoutes from App.tsx (#316 phase 8c)
The last route group, and the largest: everything reached from inside a run.
Node map, the node-type screens (event, merchant, mystery, camp, memory,
character encounter, narrator journal, item found), the between-node flow
(reward, cutscene, boss dialogue, boss epilogue), replay briefing, relic
select, card rest, starter pack, and the end-of-act/end-of-campaign screens.

With this, App.tsx renders no screens at all. Its return is 22 lines — the
container, the two providers, Suspense, six route groups and AppOverlays —
down from a single 1,156-line block with ~85 screen branches. The lazyScreens
import in App is gone entirely; every code-split screen is now imported by the
route group that renders it, and the emitted chunks are unchanged.

AppContextValue gains the campaign run state, the four flow refs
(cutsceneDoneRef, epilogueDoneRef, replayBriefingRef, relicSelectDoneRef,
brokenRelicRef) and the campaign handlers.

App.tsx: 2,894 -> 2,684 lines (4,205 at the start of this PR). Build green,
chunk boundaries unchanged, 2107 tests passing.
